WIP: set up external ingestion flow. pending confirmation
