+++ src/service/pipeline/mod.rs
@@ -26,6 +26,7 @@ use crate::common::{
 };
 
 pub mod batch_execution;
+mod external_ingestion;
 
 #[tracing::instrument(skip(pipeline))]
 pub async fn save_pipeline(mut pipeline: Pipeline) -> Result<(), PipelineError> {
